# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API OMPAYE",
        "description": "API pour la gestion des comptes et transactions OMPAYE",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement local"
        },
        {
            "url": "https://ompaye-api.onrender.com",
            "description": "Serveur de production"
        },
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/v1/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Lister les transactions de l'utilisateur",
                "description": "Récupérer toutes les transactions de l'utilisateur connecté",
                "operationId": "getTransactions",
                "responses": {
                    "200": {
                        "description": "Liste des transactions récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Transaction"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/transactions/depot": {
            "post": {
                "tags": [
                    "Transactions Distributeur"
                ],
                "summary": "Effectuer un dépôt (Distributeur uniquement)",
                "description": "Créditer le compte d'un client avec un montant spécifié - Réservé aux distributeurs",
                "operationId": "depot",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "numero_compte",
                                    "montant"
                                ],
                                "properties": {
                                    "numero_compte": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "montant": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0.01,
                                        "example": 100.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dépôt effectué avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dépôt effectué avec succès."
                                        },
                                        "data": {
                                            "properties": {
                                                "transaction": {
                                                    "$ref": "#/components/schemas/Transaction"
                                                },
                                                "solde_actuel": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 1100.5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé - Réservé aux distributeurs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Accès refusé. Cette fonctionnalité est réservée aux distributeurs."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The numero_compte field is required."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Transaction": {
                "title": "Transaction",
                "description": "Modèle représentant une transaction",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "compte_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "merchant_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "depot",
                            "retrait",
                            "paiement",
                            "transfert"
                        ],
                        "example": "depot"
                    },
                    "montant": {
                        "type": "number",
                        "format": "float",
                        "example": 100.5
                    },
                    "statut": {
                        "type": "string",
                        "enum": [
                            "en_attente",
                            "valide",
                            "annule"
                        ],
                        "example": "valide"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrez le token JWT dans le format: Bearer {token}"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}